+++ src/BuildingBlocks/Common/Authorization/Permissions.cs
@@ -9,5 +9,12 @@ public static class Permissions
         public const string Delete = "permissions:delete";
         public const string Update = "permissions:update";
     }
+    public static class Payment
+    {
+        public const string Read = "payment:read";
+        public const string Write = "payment:write";
+        public const string Delete = "payment:delete";
+        public const string Update = "payment:update";
+    }
     
 }