+++ src/Services/Identity/Identity.API/Program.cs
@@ -32,9 +32,9 @@ try
     app.Run();
 
 }
-catch (Exception ex)
+catch (Exception ex) when (ex is not HostAbortedException && ex.Source != "Microsoft.EntityFrameworkCore.Design")  // see https://github.com/dotnet/efcore/issues/29923
 {
-    Log.Fatal(ex.Source, "Identity Service start-up failed");
+    Log.Fatal(ex, "Identity Service start-up failed");
 }
 finally
 {