+++ src/Services/PaymentOrchestration/PaymentOrchestration.API/Extensions/ApplicationBuilderExtensions.cs
@@ -1,5 +1,8 @@
-﻿using Serilog;
+﻿using System.Globalization;
+using Common.Extensions.Middlewares;
+using Serilog;
 using Infrastructure.Extensions.Middlewares;
+using Prometheus;
 
 namespace PaymentOrchestration.API.Extensions;
 
@@ -9,25 +12,50 @@ public static class ApplicationBuilderExtensions
     {
         app.UseHttpsRedirection();
 
+        app.UseRouting();
+        app.UseAuthentication();
+        app.UseAuthorization();
+
+        app.UseLocalizationMiddlewares();
         app.UseInfrastructureMiddlewares();
 
-        app.UseSecurityMiddlewares();
-        
         app.MapControllers();
-        
+
+        return app;
+    }
+    
+    private static WebApplication UseLocalizationMiddlewares(this WebApplication app)
+    {
+        var supportedCultures = new[]
+        {
+            new CultureInfo("en-US"),
+            new CultureInfo("tr-TR")
+        };
+
+        var localizationOptions = new RequestLocalizationOptions
+        {
+            DefaultRequestCulture = new Microsoft.AspNetCore.Localization.RequestCulture("en-US"),
+            SupportedCultures = supportedCultures,
+            SupportedUICultures = supportedCultures
+        };
+
+        app.UseRequestLocalization(localizationOptions);
         return app;
     }
 
     private static WebApplication UseInfrastructureMiddlewares(this WebApplication app)
     {
+        app.UseHttpMetrics();
+        app.MapMetrics();
         app.UseSerilogRequestLogging();
         app.UseCorrelationId();
+        app.UseGlobalExceptionHandlerMiddleware();
         return app;
     }
     
     private static WebApplication UseSecurityMiddlewares(this WebApplication app)
     {
-        app.UseAuthorization();
+        app.UseClaimsFromHeadersMiddleware();
         return app;
     }
 }