+++ src/Services/PaymentOrchestration/PaymentOrchestration.Application/AssemblyReference.cs
@@ -1,6 +1,3 @@
 ﻿namespace PaymentOrchestration.Application;
 
-public class Class1
-{
-
-}
+public static class AssemblyReference {}