+++ src/Services/PaymentOrchestration/PaymentOrchestration.Application/Features/Commands/CreatePaymentRequest/CreatePaymentRequestCommand.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace PaymentOrchestration.Application.Features.Commands.CreatePaymentRequest;
+
+public record CreatePaymentRequestCommand(
+    decimal Amount,
+    string Currency,
+    string CardHolderName,
+    string TokenizedCardNumber) : IRequest<Guid>;