+++ src/Services/PaymentOrchestration/PaymentOrchestration.Domain/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace PaymentOrchestration.Domain;
-
-public class Class1
-{
-
-}