+++ src/Services/PaymentOrchestration/PaymentOrchestration.Domain/Common/Abstract/IAggregateRoot.cs
@@ -1,0 +1,8 @@
+﻿namespace PaymentOrchestration.Domain.Common.Abstract;
+
+public interface IAggregateRoot
+{
+    IReadOnlyCollection<IDomainEvent> DomainEvents { get; }
+    void AddDomainEvent(IDomainEvent domainEvent);
+    void ClearDomainEvents();
+}