+++ src/Services/PaymentOrchestration/PaymentOrchestration.Infrastructure/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace PaymentOrchestration.Infrastructure;
-
-public class Class1
-{
-
-}