+++ src/Gateways/ApiGateway/CHANGELOG.md
@@ -1,8 +1,26 @@
 ﻿# Changelog
 
-## Latest Version : v0.1.0
+## Latest Version : v0.2.0
 
 ## [Unreleased]
+
+## [v0.2.0] - 2025-08-21
+
+### Added
+
+- JWT authentication has been added.
+
+### Changed
+
+- No changes.
+
+### Fixed
+
+- No changes.
+
+### Removed
+
+- No changes.
 
 ## [v0.1.0] - 2025-07-25
 
@@ -14,4 +32,12 @@
 
 ### Changed
 
+- No changes.
+
 ### Fixed
+
+- No changes.
+
+### Removed
+
+- No changes.