+++ src/Services/PaymentOrchestration/CHANGELOG.md
@@ -1,8 +1,46 @@
 # Changelog
 
-## Latest Version : v0.1.0
+## Latest Version : v0.2.0
 
 ## [Unreleased]
+
+## [v0.2.0] - 2025-08-21
+
+### Added
+
+- Payment Controller has been codded.
+- GetPaymentById and CreatePayment endpoints have been codded.
+- CQRS Pattern implemented.
+- CreatePaymentRequestCommand has been added.
+- CreatePaymentRequestCommandHandler has been added.
+- GetPaymentRequestByIdQuery has been added.
+- GetPaymentRequestByIdQueryHandler has been added.
+- JWT Authentication has been added.
+- Role Based Authorization has been added.
+- Redis blacklist token check middleware has been added.
+- OpenTelemetry + Jaeger have been configured.
+- Localization has been added.
+- CorrelationId, GlobalExceptionHandler, Serilog, Prometheus have been configured.
+- Language resources have been added.
+- Fluent Validation and Validators have been added.
+- PaymentRequestSagaState has been coded.
+- PaymentRequestSagaStateMachine has been coded.
+- Payment Aggreate has been coded.
+- Repository pattern implemented.
+- PaymentDbContext has been coded.
+-
+
+### Changed
+
+- No changes.
+
+### Fixed
+
+- No changes.
+
+### Removed
+
+- No changes.
 
 ## [v0.1.0] - 2025-07-19